finish settings, remove reset settings.

+++ HebrewCalendar/SettingsView.qml
@@ -7,24 +7,7 @@ Page {
     id: settingsViewPage
     objectName: "settingsViewPage"
     property bool isCurrentPage: false
-    Action {
-        id: settingsReset
-        objectName:"settingsReset"
-        iconName: "reset"
-        text: i18n.tr("Reset")
-        onTriggered: {
 
-        }
-    }
-    head {
-        actions: [
-            settingsReset
-        ]
-        contents: Label {
-            fontSize: "x-large"
-            text: i18n.tr("Settings")
-        }
-    }
     Settings{
         id:settings
     }
@@ -52,6 +35,36 @@ Page {
                     "longitude":["34.77"],
                     "timeZone":["Asia/Tel_Aviv"]}
     }
+    U1db.Document {
+        id: locationTemplate3
+        database: loacationDatabase
+        docId: 'firstRun3'
+        create: true
+        defaults: { "name":["Beher-Sheva"],
+                    "latitude":["31.24"],
+                    "longitude":["34.79"],
+                    "timeZone":["Asia/Tel_Aviv"]}
+    }
+    U1db.Document {
+        id: locationTemplate4
+        database: loacationDatabase
+        docId: 'firstRun4'
+        create: true
+        defaults: { "name":["Haifa"],
+                    "latitude":["32.8"],
+                    "longitude":["34.99"],
+                    "timeZone":["Asia/Tel_Aviv"]}
+    }
+    U1db.Document {
+        id: locationTemplate5
+        database: loacationDatabase
+        docId: 'firstRun5'
+        create: true
+        defaults: { "name":["Bnei-Brak"],
+                    "latitude":["32.08"],
+                    "longitude":["34.82"],
+                    "timeZone":["Asia/Tel_Aviv"]}
+    }
     U1db.Index{
         database: loacationDatabase
         id: loacationIndex
@@ -64,7 +77,7 @@ Page {
 
     Column {
         anchors.fill: parent
-        anchors.margins:  units.gu(3)
+        anchors.margins: units.gu(2)
         spacing: units.gu(2)
         ListItem.Header  {
             text: i18n.tr("Global")
@@ -118,47 +131,48 @@ Page {
                     locationSelector.selectedIndex=getIndexOf(select);
                 }
                 var selectedIndex = locationSelector.selectedIndex
-                var m = model;
-                var s =locationSelector.selectedIndex;
+                if(selectedIndex>=locationSelector.model.results.length){
+                    selectedIndex=0;
+                }
                 settings.locationName =locationSelector.model.results[selectedIndex].name[0]
                 settings.locationLongitude=locationSelector.model.results[selectedIndex].longitude[0]
                 settings.locationLatitude=locationSelector.model.results[selectedIndex].latitude[0]
                 settings.locationTimeZone=locationSelector.model.results[selectedIndex].timeZone[0]
-                console.log(settings.locationName+" "+settings.locationLongitude+" "+settings.locationLatitude+" "
-                            +settings.locationTimeZone)
             }
         }
 
         Row {
-            anchors.left: parent.left
-            anchors.right: parent.right
+            anchors.horizontalCenter:parent.horizontalCenter
             spacing: units.gu(2)
+
             Button {
                 iconName: "edit"
-              //  text: i18n.tr("Edit location")
+                width: units.gu(10)
                 onClicked:{
                     PopupUtils.open(Qt.resolvedUrl("LocationDialog.qml"),parent, {
                                         locationDocId:locationSelector.model.documents[locationSelector.selectedIndex],
                                         locationName: locationSelector.model.results[locationSelector.selectedIndex].name[0],
                                         locationLongitude:locationSelector.model.results[locationSelector.selectedIndex].longitude[0],
-                                        locationLatitude:locationSelector.model.results[locationSelector.selectedIndex].longitude[0],
+                                        locationLatitude:locationSelector.model.results[locationSelector.selectedIndex].latitude[0],
                                         locationTimeZone:locationSelector.model.results[locationSelector.selectedIndex].timeZone[0]})
 
                 }
             }
             Button {
                 iconName: "add"
-               // text: i18n.tr("New loaction")
+                width: units.gu(10)
                 onClicked:{
                     PopupUtils.open(Qt.resolvedUrl("LocationDialog.qml"),parent)
                 }
             }
             Button {
+                 width: units.gu(10)
                 iconName: "delete"
-                //text: i18n.tr("Delete loaction")
                 onClicked:{
+                    if(locationSelector.model.documents.length>1){
                         loacationDatabase.deleteDoc(locationSelector.model.documents[locationSelector.selectedIndex]);
                         locationSelector.updateSettings("")
+                    }
                 }
             }
         }
